app/requirements.txt
gradio==5.1.0
mysql-connector-python==8.4.0
sqlalchemy==2.0.31
openai==1.35.14